+++ src/main/java/com/treasure/hunt/strategy/hider/implementations/CircleTipster.java
@@ -1,13 +1,13 @@
-package com.treasure.hunt.strategy.tipster.implementations;
+package com.treasure.hunt.strategy.hider.implementations;
 
 import com.treasure.hunt.game.GameHistory;
 import com.treasure.hunt.strategy.Product;
+import com.treasure.hunt.strategy.hider.Hider;
 import com.treasure.hunt.strategy.hint.CircleHint;
-import com.treasure.hunt.strategy.seeker.Moves;
-import com.treasure.hunt.strategy.tipster.Tipster;
+import com.treasure.hunt.strategy.searcher.Moves;
 import org.locationtech.jts.geom.Point;
 
-public class CircleTipster implements Tipster<CircleHint> {
+public class CircleTipster implements Hider<CircleHint> {
     @Override
     public void init(Point treasurePosition, GameHistory gameHistory) {
