+++ src/main/java/com/treasure/hunt/strategy/searcher/implementations/CircleSeeker.java
@@ -1,13 +1,13 @@
-package com.treasure.hunt.strategy.seeker.implementations;
+package com.treasure.hunt.strategy.searcher.implementations;
 
 import com.treasure.hunt.game.GameHistory;
 import com.treasure.hunt.strategy.Product;
 import com.treasure.hunt.strategy.hint.CircleHint;
-import com.treasure.hunt.strategy.seeker.Moves;
-import com.treasure.hunt.strategy.seeker.Seeker;
+import com.treasure.hunt.strategy.searcher.Moves;
+import com.treasure.hunt.strategy.searcher.Searcher;
 import org.locationtech.jts.geom.Point;
 
-public class CircleSeeker implements Seeker<CircleHint> {
+public class CircleSeeker implements Searcher<CircleHint> {
 
     @Override
     public void init(Point startPosition, GameHistory gameHistory) {
